add 3 callback to MatterInterfaceHandler, but network call back not work correctly. so comment it.
