+++ .vscode/settings.json
@@ -6,6 +6,12 @@
     "idf.flashType": "UART",
     "idf.port": "/dev/ttyACM0",
     "files.associations": {
-        "string": "cpp"
+        "string": "cpp",
+        "array": "cpp",
+        "compare": "cpp",
+        "functional": "cpp",
+        "tuple": "cpp",
+        "type_traits": "cpp",
+        "utility": "cpp"
     }
 }
