+++ components/BasicTemplates/DerivedProcess.cpp
@@ -6,28 +6,28 @@
 
 namespace DoneMatter {
 
-void DerivedProcess::checkEssencialParam()
-{
-    // if (InterfaceHandler.SharedBufQueue != NULL &&
-    //     InterfaceHandler.SharedSemaphore != NULL &&
-    //     InterfaceHandler.localBufQueue != NULL &&
-    //     InterfaceHandler.localSemaphore != NULL)        
-    // {
-    //     return true
-    // }
-    // else
-    //     return false;
-    return true;
-}
+// bool DerivedProcess::checkEssencialParam()
+// {
+//     // if (InterfaceHandler.SharedBufQueue != NULL &&
+//     //     InterfaceHandler.SharedSemaphore != NULL &&
+//     //     InterfaceHandler.localBufQueue != NULL &&
+//     //     InterfaceHandler.localSemaphore != NULL)        
+//     // {
+//     //     return true
+//     // }
+//     // else
+//     //     return false;
+//     return true;
+// }
 
-void DerivedProcess::Derived_MainTask(void *pvParameters)
-{
-    ESP_LOGI("MainTask","Derived_MainTask1Time");
+// void DerivedProcess::Derived_MainTask(void *pvParameters)
+// {
+//     ESP_LOGI("MainTask","Derived_MainTask1Time");
 
-    while (true)
-    {
-        vTaskDelay(pdMS_TO_TICKS(taskDelay));
-        ESP_LOGI("MainTask","Derived_MainTask4ever");
-    }    
-}
+//     while (true)
+//     {
+//         //vTaskDelay(pdMS_TO_TICKS(taskDelay));
+//         ESP_LOGI("MainTask","Derived_MainTask4ever");
+//     }    
+// }
 }
