+++ components/BasicTemplates/DerivedProcess.h
@@ -14,15 +14,15 @@ namespace DoneMatter {
             SemaphoreHandle_t localSemaphore;          
     };    
 
-    class DerivedProcess : public BaseProcess
-    {
-        private:
-            static void Derived_MainTask(void *pvParameters);    
+    // class DerivedProcess : public BaseProcess
+    // {
+    //     private:
+    //         static void Derived_MainTask(void *pvParameters);    
 
-        public:
-            DerivedProcess(uint32_t td) {MainTask = Derived_MainTask;};
-            ~DerivedProcess(){};
+    //     public:
+    //         DerivedProcess(uint32_t td) {MainTask = Derived_MainTask;};
+    //         ~DerivedProcess(){};
         
-    };           
+    // };           
 }
 #endif //DERIVED_PROCESS_H
