--- components/MatterComponent/MatterComponent.h
@@ -1,10 +1,0 @@
-#ifndef MATTER_INTERFACE_H
-#define MATTER_INTERFACE_H
-
-#include "freertos/FreeRTOS.h"
-#include "freertos/task.h"
-#include "freertos/semphr.h"
-
-void matterComp(void);
-
-#endif //MATTER_INTERFACE_H
